feat(home): add TopPlayerCard component and related logic
- Implemented TopPlayerCard component for displaying player run details.
- Created homeIcons for various UI elements including Search, Calendar, Timer, and platform icons.
- Developed homeLogic to manage filtering and selection of runs, including utility functions for handling run data.
- Introduced homeMedia for character image handling with fallback mechanisms.
- Defined types for HomeRun, DisplayBucket, and filter states to enhance type safety and clarity in the codebase.
